5328: Refactor RefreshMaterializedViews
