+++ src/server/db/repository/assessmentCycle/countryActivityLog/index.ts
@@ -1,6 +1,7 @@
 import { createMaterializedView } from 'server/db/repository/assessmentCycle/countryActivityLog/createMaterializedView'
 import { dropMaterializedView } from 'server/db/repository/assessmentCycle/countryActivityLog/dropMaterializedView'
 import { getCount } from 'server/db/repository/assessmentCycle/countryActivityLog/getCount'
+import { getCountryISOsOutOfSync } from 'server/db/repository/assessmentCycle/countryActivityLog/getCountryISOsOutOfSync'
 import { getMany } from 'server/db/repository/assessmentCycle/countryActivityLog/getMany'
 import { refreshMaterializedView } from 'server/db/repository/assessmentCycle/countryActivityLog/refreshMaterializedView'
 
@@ -8,6 +9,7 @@ export const CountryActivityLogRepository = {
   createMaterializedView,
   dropMaterializedView,
   getCount,
+  getCountryISOsOutOfSync,
   getMany,
   refreshMaterializedView,
 }
