+++ src/server/worker/jobs/refreshMaterializedViews.ts
@@ -1,76 +1,34 @@
 import { Promises } from 'utils/promises'
 
-import { CountryIso } from 'meta/area/countryIso'
-import { Assessment } from 'meta/assessment/assessment'
-import { Cycle } from 'meta/assessment/cycle'
-
 import { AreaController } from 'server/controller/area'
 import { AssessmentController } from 'server/controller/assessment'
 import { BaseProtocol, DB } from 'server/db/db'
 import { CountryActivityLogRepository } from 'server/db/repository/assessmentCycle/countryActivityLog'
-import {
-  activitiesLastEdit,
-  activitiesLastEditOdpData,
-} from 'server/db/repository/assessmentCycle/countrySummary/_lastEditActivities'
-import { Schemas } from 'server/db/schemas'
 import { Logger } from 'server/utils/logger'
+import { Job } from 'server/worker/jobs/job'
 
 const client: BaseProtocol = DB
 const name = 'Scheduler-MaterializedViews'
 
-const _getCountryISOsOutOfSync = async (props: {
-  assessment: Assessment
-  cycle: Cycle
-}): Promise<Array<CountryIso>> => {
-  const { assessment, cycle } = props
+export class RefreshMaterializedViews extends Job {
+  constructor() {
+    super(name)
+  }
 
-  const schemaCycle = Schemas.getNameCycle(assessment, cycle)
-  const countries = await AreaController.getCountries({ assessment, cycle })
-  const query = `with al as
-    (select al.country_iso
-          , max(al.time) filter (where al.message in (${activitiesLastEdit}))         as last_edit
-                  , max(al.time) filter (where al.message in (${activitiesLastEditOdpData})) as last_odp_edit
-     from activity_log al
-     where al.assessment_uuid = $1
-       and al.cycle_uuid = $2
-     group by al.country_iso)
-                    , al_country as (
-                   ${countries.map((country) => {
-                     const { countryIso } = country
-                     return `select '${countryIso}' as country_iso, max(time) as last_edit
-                             from ${schemaCycle}."activity_log_${countryIso}"
-                             where message in (${activitiesLastEdit})`
-                   }).join(`
-    union
-    `)})
-  select c.country_iso as countryiso
-  from ${schemaCycle}.country c
-         left join al on c.country_iso = al.country_iso
-         left join al_country alc on c.country_iso = alc.country_iso and al.country_iso = alc.country_iso
-  where (alc.last_edit is null and (al.last_edit is not null or al.last_odp_edit is not null))
-     or greatest(al.last_edit, al.last_odp_edit) > alc.last_edit
-  order by 1 desc`
-  const res = await client.map(query, [assessment.uuid, cycle.uuid], ({ countryiso }) => countryiso)
+  protected async execute(): Promise<void> {
+    const assessments = await AssessmentController.getAll({}, client)
 
-  return res
-}
+    await Promises.each(assessments, (assessment) =>
+      Promises.each(assessment.cycles, async (cycle) => {
+        const countries = await AreaController.getCountries({ assessment, cycle })
+        const countryISOs = await CountryActivityLogRepository.getCountryISOsOutOfSync({ assessment, countries, cycle })
 
-export const refreshMaterializedViews = async (): Promise<void> => {
-  Logger.info(`[${name}] ** started`)
-
-  const assessments = await AssessmentController.getAll({}, client)
-
-  await Promises.each(assessments, (assessment) =>
-    Promises.each(assessment.cycles, async (cycle) => {
-      // 1. refresh countries activity log
-      const countryISOs = await _getCountryISOsOutOfSync({ assessment, cycle })
-      await Promises.each(countryISOs, async (countryIso) => {
-        Logger.debug(`[${name}:CountryActivityLog] ${assessment.props.name} ${cycle.name} ${countryIso} refreshing`)
-        await CountryActivityLogRepository.refreshMaterializedView({ assessment, cycle, countryIso })
-        Logger.info(`[${name}:CountryActivityLog] ${assessment.props.name} ${cycle.name} ${countryIso} refreshed`)
+        await Promises.each(countryISOs, async (countryIso) => {
+          Logger.debug(`[${name}:CountryActivityLog] ${assessment.props.name} ${cycle.name} ${countryIso} refreshing`)
+          await CountryActivityLogRepository.refreshMaterializedView({ assessment, cycle, countryIso })
+          Logger.info(`[${name}:CountryActivityLog] ${assessment.props.name} ${cycle.name} ${countryIso} refreshed`)
+        })
       })
-    })
-  )
-
-  Logger.info(`[${name}] ** terminated`)
+    )
+  }
 }
